refactor by current master branch
