+++ src/test/java/school/faang/user_service/service/SubscriptionServiceTest.java
@@ -14,7 +14,7 @@ import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.exception.UserWasNotFoundException;
 import school.faang.user_service.mapper.UserFollowingMapper;
 import school.faang.user_service.mapper.UserFollowingMapperImpl;
-import school.faang.user_service.rating.publisher.FollowerEventPublisher;
+import school.faang.user_service.messaging.FollowerEventPublisher;
 import school.faang.user_service.rating.publisher.UserEventPublisher;
 import school.faang.user_service.repository.SubscriptionRepository;
 import school.faang.user_service.repository.UserRepository;
